move accept cookies into pagebase
