+++ ifsc/pages/home.py
@@ -6,11 +6,6 @@ from ifsc.pages.news import News
 class Home(PageBase):
     home_by = (By.XPATH, "//nav/div[2]/ul/li[1]/a")
     news_by = (By.XPATH, "//nav/div[2]/ul/li[6]/a")
-    dialog_div = (By.ID, "onetrust-banner-sdk")
-    accept_all_cookies_by = (
-        By.XPATH,
-        "/html/body/div[2]/div[3]/div/div[1]/div/div[2]/div/button[3]",
-    )
 
     def __init__(self):
         super().__init__()
@@ -28,14 +23,3 @@ class Home(PageBase):
         news_link.click()
         news = News()
         return news
-
-    def clickAcceptAllCookies(self):
-        accept_button = self.waits.waitForElementToBeClickable(
-            self.accept_all_cookies_by
-        )
-        accept_button.click()
-        return self
-
-    def waitCookiesDialogClose(self):
-        dialog_div = self.waits.getInvisibilityOfElementLocated(self.dialog_div)
-        return self
